+++ iniesta/listeners.py
@@ -1,3 +1,4 @@
+from insanic.conf import settings
 from insanic.log import error_logger
 
 from iniesta.sns import SNSClient
@@ -6,14 +7,13 @@ from iniesta.sqs import SQSClient
 
 class IniestaListener:
 
-    def __init__(self, sns_endpoint_url=None, sqs_endpoint_url=None):
-        self.sns_endpoint_url = sns_endpoint_url
-        self.sqs_endpoint_url = sqs_endpoint_url
+    def __init__(self):
+        self.sns_endpoint_url = settings.INIESTA_SNS_ENDPOINT_URL
+        self.sqs_endpoint_url = settings.INIESTA_SQS_ENDPOINT_URL
 
     async def _initialize_sns(self, app):
         app.xavi = await SNSClient.initialize(
-            topic_arn=app.config.INIESTA_SNS_PRODUCER_GLOBAL_TOPIC_ARN,
-            endpoint_url=self.sns_endpoint_url
+            topic_arn=app.config.INIESTA_SNS_PRODUCER_GLOBAL_TOPIC_ARN
         )
 
     async def _initialize_sqs(self, app):
@@ -22,8 +22,7 @@ class IniestaListener:
             queue_name=app.config.INIESTA_SQS_QUEUE_NAME_TEMPLATE.format(
                 env=app.config.MMT_ENV,
                 service_name=app.config.SERVICE_NAME
-            ),
-            endpoint_url=self.sqs_endpoint_url
+            )
         )
 
     def _start_polling(self, app):
@@ -32,9 +31,7 @@ class IniestaListener:
     async def _stop_polling(self, app):
         await app.messi.stop_receiving_messages()
 
-
     # actual listeners
-
     async def after_server_start_producer_check(self, app, loop=None, **kwargs):
         await self._initialize_sns(app)
 
@@ -56,7 +53,7 @@ class IniestaListener:
 
         await app.messi.confirm_subscription(
             app.config.INIESTA_SNS_PRODUCER_GLOBAL_TOPIC_ARN,
-            sns_endpoint_url=self.sns_endpoint_url
+
         )
         await app.messi.confirm_permission(
             app.config.INIESTA_SNS_PRODUCER_GLOBAL_TOPIC_ARN