+++ HISTORY.rst
@@ -8,7 +8,8 @@ History
 
 - FEAT: separate initialization methods for purposes
 - FEAT: allow decorating of default(fallback) handler
-
+- CHORE: more test coverage
+- CHORE: refactor endpoint urls to settings
 
 0.1.1 (2019-03-22)
 ++++++++++++++++++